.history/Rag/requirements_20240916205306.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
langchain
langchain_community
openai
python-dotenv
qdrant-client
tiktoken
pandas
langchain-openai
langchain-qdrant
azure-storage-blob
azure-identity
mysql.connector